Arreglo del flag x en modo diálogo
Antes no se escribía la regla si el flag x estaba activo pero el d no. Aparte de esto, se podía producir un overflow en los mensajes, ahora está manejado.

También he añadido algo de documentación y he cambiado algunos detalles.

+++ Operaciones/Calculos.cs
@@ -154,7 +154,7 @@ namespace Operaciones {
 		 * @param pos cifra que calcular
 		 * @return cifra {@code pos} de {@code num} en la raiz del objeto
 		 */
-		public byte Cifra(long num, long pos)
+		public short Cifra(long num, long pos)
 		{
 			return CalculosEstatico.Cifra(num, pos, _raiz);
 		}
